Add standardized font sizes
Adds standardized font sizes for APs and assistants. These are defined
inside values/dimens.xml. The intention is that all font sizes in the
app are defined in terms of these. There are two standards: one for APs
and one for assistants. The idea is that screens/text that is intended
for APs utilize the AP standard, and text for assistants will utilize
the assistant standard.
Also fix bug where you could log in despite a short password.