Add helper button to AP Map activity and more
A helper button has been added to the AP Map Activity, replacing the
existing Info button and Settings button.
The floating "New Task" button on the AP Map Activity has been removed in
favor of a dynamic "Task Button" at the bottom of the screen. This
operates under the new constraint of APs only being able to have one
active task at a time. If a task has been created, the Task Button will
operate as a "view current task" button, while if no task exists, it acts
as a new task button. The "view current task" button currently just
redirects the user to the login screen - this will be changed when the
actual screen is implemented.
The task confirm button now saves the ImmutableTask object inside
ClientInfo, which now has additional member and helper methods to
facilitate this.
Added more strings to strings.xml.
This change includes tests and changes to existing tests.