Add text chat window and functionality
Currently placed in the top right of the assistant home map activity.
However, it allows full one to one chat. Simply needs to be dropped in and
integrated once a proper user activity has been added.
Added a few extra dependencies to fix compile issues, and had to increase
min sdk to 16.
Made some changes to FirebaseAdapter to allow for the new changes.
Also includes test for some of the added code. However, I was unable to
make any tests for the chat activity, and will need to consult Oscar on
how to do them, due to onCreate running before I want it to.
All tests pass in this commit.