Fix a few bugs and more
Fixed a bug where the text on the Task Buttons wouldn't properly update
to reflect whether the user had a task. This has been fixed by moving
that code into the 'onResume' method.
Updated gradle to 3.2.
Now, when an assistant accepts a task, rather than going back to the
task list, they're brought to their home screen.
Fixed a bug in the TaskCurrent screen in which the app would crash when
the task assistant was null.