Fix a few bugs
Fix a bug where logging into an account with an in-progress task would
not be registered i.e. the app would not check Firebase and thus display
"new task" button for both assistants and APs.
Fix the chat screen not properly displaying colors for the
sender/receiver.
Utilize strings.xml in login screen.